requirements.txt
instacapture
pytz
requests
lxml
pillow   # GUI resim işleme için
